// src/api/models/business-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { BusinessBriefModelDataByType } from './business-brief-model-data-by-type';
// May contain unused imports in some cases
// @ts-ignore
import { BusinessIdentificationType } from './business-identification-type';
// May contain unused imports in some cases
// @ts-ignore
import { BusinessType } from './business-type';
// May contain unused imports in some cases
// @ts-ignore
import { BusinessUserModel } from './business-user-model';
// May contain unused imports in some cases
// @ts-ignore
import { ContactInfoModel } from './contact-info-model';
// May contain unused imports in some cases
// @ts-ignore
import { RoleProfileModel } from './role-profile-model';
// May contain unused imports in some cases
// @ts-ignore
import { Status } from './status';
// May contain unused imports in some cases
// @ts-ignore
import { UserModel } from './user-model';

/**
 * 
 * @export
 * @interface BusinessModel
 */
export interface BusinessModel {
    /**
     * 
     * @type {number}
     * @memberof BusinessModel
     */
    'id'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof BusinessModel
     */
    'isNullable'?: boolean;
    /**
     * 
     * @type {Status}
     * @memberof BusinessModel
     */
    'status'?: Status;
    /**
     * 
     * @type {string}
     * @memberof BusinessModel
     */
    'createdBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessModel
     */
    'createdOnUtc'?: string;
    /**
     * 
     * @type {string}
     * @memberof BusinessModel
     */
    'updatedBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessModel
     */
    'updatedOnUtc'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessModel
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessModel
     */
    'code'?: string | null;
    /**
     * 
     * @type {BusinessIdentificationType}
     * @memberof BusinessModel
     */
    'identificationType'?: BusinessIdentificationType;
    /**
     * 
     * @type {string}
     * @memberof BusinessModel
     */
    'identification'?: string | null;
    /**
     * 
     * @type {BusinessType}
     * @memberof BusinessModel
     */
    'type'?: BusinessType;
    /**
     * 
     * @type {BusinessBriefModelDataByType}
     * @memberof BusinessModel
     */
    'dataByType'?: BusinessBriefModelDataByType;
    /**
     * 
     * @type {number}
     * @memberof BusinessModel
     */
    'ownerId'?: number | null;
    /**
     * 
     * @type {BusinessModel}
     * @memberof BusinessModel
     */
    'owner'?: BusinessModel;
    /**
     * 
     * @type {string}
     * @memberof BusinessModel
     */
    'responsibleUserId'?: string | null;
    /**
     * 
     * @type {UserModel}
     * @memberof BusinessModel
     */
    'responsibleUser'?: UserModel;
    /**
     * 
     * @type {Array<ContactInfoModel>}
     * @memberof BusinessModel
     */
    'contactInfo'?: Array<ContactInfoModel> | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessModel
     */
    'addressLine1'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessModel
     */
    'addressLine2'?: string | null;
    /**
     * 
     * @type {Array<BusinessUserModel>}
     * @memberof BusinessModel
     */
    'users'?: Array<BusinessUserModel> | null;
    /**
     * 
     * @type {Array<RoleProfileModel>}
     * @memberof BusinessModel
     */
    'profiles'?: Array<RoleProfileModel> | null;
}