// src/api/models/role-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { Status } from './status';

/**
 * 
 * @export
 * @interface RoleModel
 */
export interface RoleModel {
    /**
     * 
     * @type {string}
     * @memberof RoleModel
     */
    'id'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof RoleModel
     */
    'isNullable'?: boolean;
    /**
     * 
     * @type {Status}
     * @memberof RoleModel
     */
    'status'?: Status;
    /**
     * 
     * @type {string}
     * @memberof RoleModel
     */
    'createdBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RoleModel
     */
    'createdOnUtc'?: string;
    /**
     * 
     * @type {string}
     * @memberof RoleModel
     */
    'updatedBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RoleModel
     */
    'updatedOnUtc'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RoleModel
     */
    'name'?: string | null;
}