// src/api/models/role-profile-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { BusinessModel } from './business-model';
// May contain unused imports in some cases
// @ts-ignore
import { RoleModel } from './role-model';
// May contain unused imports in some cases
// @ts-ignore
import { Status } from './status';

/**
 * 
 * @export
 * @interface RoleProfileModel
 */
export interface RoleProfileModel {
    /**
     * 
     * @type {number}
     * @memberof RoleProfileModel
     */
    'id'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof RoleProfileModel
     */
    'isNullable'?: boolean;
    /**
     * 
     * @type {Status}
     * @memberof RoleProfileModel
     */
    'status'?: Status;
    /**
     * 
     * @type {string}
     * @memberof RoleProfileModel
     */
    'createdBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RoleProfileModel
     */
    'createdOnUtc'?: string;
    /**
     * 
     * @type {string}
     * @memberof RoleProfileModel
     */
    'updatedBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RoleProfileModel
     */
    'updatedOnUtc'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RoleProfileModel
     */
    'code'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RoleProfileModel
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RoleProfileModel
     */
    'description'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof RoleProfileModel
     */
    'businessId'?: number | null;
    /**
     * 
     * @type {BusinessModel}
     * @memberof RoleProfileModel
     */
    'business'?: BusinessModel;
    /**
     * 
     * @type {Array<RoleModel>}
     * @memberof RoleProfileModel
     */
    'roles'?: Array<RoleModel> | null;
    /**
     * 
     * @type {boolean}
     * @memberof RoleProfileModel
     */
    'isDefault'?: boolean;
}