// src/api/models/user-model.ts
export interface UserModel {
    /**
     * 
     * @type {string}
     * @memberof UserModel
     */
    'id'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof UserModel
     */
    'isNullable'?: boolean;
    /**
     * 
     * @type {Status}
     * @memberof UserModel
     */
    'status'?: Status;
    /**
     * 
     * @type {string}
     * @memberof UserModel
     */
    'createdBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserModel
     */
    'createdOnUtc'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserModel
     */
    'updatedBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserModel
     */
    'updatedOnUtc'?: string | null;
    /**
     * 
     * @type {UserType}
     * @memberof UserModel
     */
    'type'?: UserType;
    /**
     * 
     * @type {string}
     * @memberof UserModel
     */
    'firstName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserModel
     */
    'secondName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserModel
     */
    'firstLastName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserModel
     */
    'secondLastName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserModel
     */
    'fullName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserModel
     */
    'phoneNumber'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof UserModel
     */
    'phoneNumberConfirmed'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof UserModel
     */
    'email'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof UserModel
     */
    'emailConfirmed'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof UserModel
     */
    'userName'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof UserModel
     */
    'lockoutEnabled'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof UserModel
     */
    'lockoutEnd'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof UserModel
     */
    'accessFailedCount'?: number;
    /**
     * 
     * @type {Array<RoleModel>}
     * @memberof UserModel
     */
    'roles'?: Array<RoleModel> | null;
    /**
     * 
     * @type {Array<RoleBriefModel>}
     * @memberof UserModel
     */
    'briefRoles'?: Array<RoleBriefModel> | null;
    /**
     * 
     * @type {Array<RoleProfileModel>}
     * @memberof UserModel
     */
    'profiles'?: Array<RoleProfileModel> | null;
    /**
     * 
     * @type {Array<RoleProfileBriefModel>}
     * @memberof UserModel
     */
    'briefProfiles'?: Array<RoleProfileBriefModel> | null;
    /**
     * 
     * @type {Array<BusinessUserBriefModel>}
     * @memberof UserModel
     */
    'briefBusinessUsers'?: Array<BusinessUserBriefModel> | null;
}